Remove: remove unused functions and field
Removed old unused functions and field in context
of the endpoint-processing of the api and from the
header of the image-datasets.
